Created 'RpcClient' struct
