+++ starknet-replay/src/runner/replay_state_reader.rs
@@ -11,13 +11,14 @@ use starknet_core::types::{ContractClass as StarknetContractClass, Felt};
 
 use crate::block_number::BlockNumber;
 use crate::runner::replay_class_hash::ReplayClassHash;
-use crate::storage::rpc::{contract_class, RpcStorage};
+use crate::storage::rpc::contract_class;
+use crate::storage::rpc::rpc_client::RpcClient;
 
 /// This structure is used by [`blockifier`] to access blockchain data during
 /// transaction replay.
 pub struct ReplayStateReader<'a> {
     /// The reference to [`crate::storage::rpc::RpcStorage`] to make RPC calls.
-    storage: &'a RpcStorage,
+    rpc_client: &'a RpcClient,
 
     /// The block number used to query the state.
     block_number: BlockNumber,
@@ -31,9 +32,9 @@ impl ReplayStateReader<'_> {
     ///   state.
     /// - `block_number`: The block number at which state is read.
     #[must_use]
-    pub fn new(storage: &RpcStorage, block_number: BlockNumber) -> ReplayStateReader<'_> {
+    pub fn new(rpc_client: &RpcClient, block_number: BlockNumber) -> ReplayStateReader<'_> {
         ReplayStateReader {
-            storage,
+            rpc_client,
             block_number,
         }
     }
@@ -45,7 +46,7 @@ impl StateReader for ReplayStateReader<'_> {
         key: StorageKey,
     ) -> StateResult<Felt> {
         let storage_value = self
-            .storage
+            .rpc_client
             .starknet_get_storage_at(&self.block_number, &contract_address, &key)
             .map_err(|err| {
                 StateError::StateReadError(
@@ -57,7 +58,7 @@ impl StateReader for ReplayStateReader<'_> {
 
     fn get_nonce_at(&self, contract_address: ContractAddress) -> StateResult<Nonce> {
         let nonce = self
-            .storage
+            .rpc_client
             .starknet_get_nonce(&self.block_number, &contract_address)
             .map_err(|err| {
                 StateError::StateReadError(
@@ -69,7 +70,7 @@ impl StateReader for ReplayStateReader<'_> {
 
     fn get_class_hash_at(&self, contract_address: ContractAddress) -> StateResult<ClassHash> {
         let class_hash = self
-            .storage
+            .rpc_client
             .starknet_get_class_hash_at(&self.block_number, &contract_address)
             .map_err(|err| {
                 StateError::StateReadError(
@@ -88,7 +89,7 @@ impl StateReader for ReplayStateReader<'_> {
             class_hash,
         };
         let contract_class = self
-            .storage
+            .rpc_client
             .starknet_get_class(&replay_class_hash)
             .map_err(|err| {
                 StateError::StateReadError(
@@ -125,7 +126,7 @@ impl StateReader for ReplayStateReader<'_> {
             class_hash,
         };
         let contract_class = self
-            .storage
+            .rpc_client
             .starknet_get_class(&replay_class_hash)
             .map_err(|err| {
                 StateError::StateReadError(
